requirements.txt
Faker
kafka-python
psycopg2